% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjust.R
\docType{methods}
\name{adjust}
\alias{adjust}
\alias{adjust,data.frame,validator-method}
\title{Adjust records to satisfy linear (in)equality constraints}
\usage{
adjust(dat, rules, ...)

\S4method{adjust}{data.frame,validator}(dat, rules, adjust,
  weight = "uniform", tol = 0.01, maxiter = 1000L, ...)
}
\arguments{
\item{dat}{An R object carrying data}

\item{rules}{An R object carrying linear (in)equation restriction rules}

\item{...}{Currently unused.}

\item{adjust}{\code{[logical array]} indicating what fields to adjust. If left unspecified,
all variables occuring in linear rules may be altered.}

\item{weight}{The adjustment weight assigned to each variable. Variables with a higher
weight will be altered less. The following values are allowed:
\itemize{
\item{\code{"unity"} If weights are not set, all variables are treated equal (with weight 1).}
\item{\code{"ratio"} Weights are set so that the ratio between adjusted variables remain
 approximately constant.}
\item{\code{numeric} vector with the same number of variables as \code{dat}. Custom weights
applied to each record.}
\item{\code{numeric array} of the same size as \code{dat}. Custom weights for each record}
}}

\item{tol}{Convergence criterion: maximum contribution of a single value to the final deviation
from the rules.}

\item{maxiter}{Maximum number of iterations for the optimization algorithm per record
(see \code{lintools::\link[lintools]{project}})}
}
\description{
Adjust records to satisfy linear (in)equality constraints
}

