% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjust.R
\docType{package}
\name{validate.adjust}
\alias{validate.adjust}
\alias{validate.adjust-package}
\title{Adjust records to satisfy linear (in)equality constraints}
\description{
Adjust records to satisfy linear (in)equality constraints
}

